Fixed backend circular dependencies
